requirements.txt
pillow==8.2.0
tensorflow-cpu==2.5.0
keras==2.4.3